Rename files, remove unnecessary toasts
